new rule: AWS.SSM.MultipleSessions
