+++ global_helpers/panther_aws_helpers.py
@@ -4,7 +4,7 @@ import os
 from typing import Any, Dict, List
 
 import boto3
-from panther_base_helpers import pantherflow_investigation
+from panther_base_helpers import deep_get, pantherflow_investigation
 from panther_config import config
 
 
@@ -226,3 +226,38 @@ def resource_lookup(resource_id: str) -> Dict[str, Any]:
 
     # Return just the attributes of the item
     return response["Item"]["attributes"]
+
+# get actor user from correct field based on identity type
+# https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-event-reference-user-identity.html#cloudtrail-event-reference-user-identity-fields
+def get_actor_user(event):
+    # IMPORTANT: We can't use event.deep_get here because this function gets called in the UDM module.
+    user_type = deep_get(event, "userIdentity", "type")
+    if user_type == "Root":
+        actor_user = deep_get(
+            event,
+            "userIdentity",
+            "userName",
+            default=deep_get(event, "userIdentity", "accountId", default="UnknownRootUser"),
+        )
+    elif user_type in ("IAMUser", "Directory", "Unknown", "SAMLUser", "WebIdentityUser"):
+        actor_user = deep_get(event, "userIdentity", "userName", default=f"Unknown{user_type}")
+    elif user_type in ("AssumedRole", "Role", "FederatedUser"):
+        actor_user = deep_get(
+            event,
+            "userIdentity",
+            "sessionContext",
+            "sessionIssuer",
+            "userName",
+            default=f"Unknown{user_type}",
+        )
+    elif user_type == "IdentityCenterUser":
+        actor_user = deep_get(
+            event, "additionalEventData", "UserName", default=f"Unknown{user_type}"
+        )
+    elif user_type in ("AWSService", "AWSAccount"):
+        actor_user = event.get("sourceIdentity", f"Unknown{user_type}")
+    elif event.get("eventType") == "AwsServiceEvent":
+        actor_user = deep_get(event, "userIdentity", "invokedBy", default="UnknownAwsServiceEvent")
+    else:
+        actor_user = "UnknownUser"
+    return actor_user
